payments/views.py
from rest_framework import viewsets
from .models import Order, Payment
from .serializers import OrderSerializer, PaymentSerializer

class OrderViewSet(viewsets.ModelViewSet):
    queryset = Order.objects.all().prefetch_related('menu_items', 'addon_options', 'orderitem_set')
    serializer_class = OrderSerializer

class PaymentViewSet(viewsets.ModelViewSet):
    queryset = Payment.objects.all().select_related('order')
    serializer_class = PaymentSerializer
