requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-allauth==65.0.2
django-cors-headers==4.4.0
django-map-widgets==0.5.1
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
pillow==10.4.0
psycopg2==2.9.9
PyJWT==2.9.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
stripe==11.1.0
typing_extensions==4.12.2
urllib3==2.2.3
